+++ src/components/charts-generic/constants.ts
@@ -1,4 +1,5 @@
 export const LEFT_MARGIN_OFFSET = 10;
+export const LEFT_MARGIN_OFFSET_GROUPED = 40;
 export const BOTTOM_MARGIN_OFFSET = 10;
 export const MINI_CHART_WIDTH = 350;
 
@@ -7,7 +8,3 @@ export const BAR_HEIGHT = 24;
 export const BAR_HEIGHT_SMALL = 18;
 export const BAR_PADDING = 16;
 export const LABEL_PADDING = 16;
-
-// Histogram chart
-export const TOOLTIP_ARROW_HEIGHT = 8;
-export const VERTICAL_TICK_OFFSET = 19;